# go/routers.go
/*
 * OpenAPI Petstore
 *
 * This is a sample server Petstore server. For this sample, you can use the api key `special-key` to test the authorization filters.
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package petstoreserver

import (
	"encoding/json"
	"net/http"
	"os"
	"strings"

	"github.com/gin-gonic/gin"
	"gopkg.in/yaml.v3"
)

// Route is the information for every URI.
type Route struct {
	// Name is the name of this Route.
	Name string
	// Method is the string for the HTTP method. ex) GET, POST etc..
	Method string
	// Pattern is the pattern of the URI.
	Pattern string
	// HandlerFunc is the handler function of this route.
	HandlerFunc gin.HandlerFunc
}

// NewRouter returns a new router.
func NewRouter(handleFunctions ApiHandleFunctions) *gin.Engine {
	return NewRouterWithGinEngine(gin.Default(), handleFunctions)
}

// NewRouter add routes to existing gin engine.
func NewRouterWithGinEngine(router *gin.Engine, handleFunctions ApiHandleFunctions) *gin.Engine {
	registerOpenAPIDocs(router)

	for _, route := range getRoutes(handleFunctions) {
		if route.HandlerFunc == nil {
			route.HandlerFunc = DefaultHandleFunc
		}
		switch route.Method {
		case http.MethodGet:
			router.GET(route.Pattern, route.HandlerFunc)
		case http.MethodPost:
			router.POST(route.Pattern, route.HandlerFunc)
		case http.MethodPut:
			router.PUT(route.Pattern, route.HandlerFunc)
		case http.MethodPatch:
			router.PATCH(route.Pattern, route.HandlerFunc)
		case http.MethodDelete:
			router.DELETE(route.Pattern, route.HandlerFunc)
		}
	}

	return router
}

// Default handler for not yet implemented routes
func DefaultHandleFunc(c *gin.Context) {
	c.String(http.StatusNotImplemented, "501 not implemented")
}

func registerOpenAPIDocs(router *gin.Engine) {
	router.GET("/openapi.yaml", func(c *gin.Context) {
		serveOpenAPIYAML(c, "api/openapi.yaml")
	})
	router.GET("/openapi.json", func(c *gin.Context) {
		serveOpenAPIJSON(c, "api/openapi.yaml")
	})
	router.GET("/swagger", serveSwaggerUI)
	router.GET("/swagger/", serveSwaggerUI)
}

func serveOpenAPIYAML(c *gin.Context, path string) {
	if _, err := os.Stat(path); err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
	c.File(path)
}

func serveOpenAPIJSON(c *gin.Context, path string) {
	data, err := os.ReadFile(path)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
	var payload interface{}
	if err := yaml.Unmarshal(data, &payload); err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
	jsonBytes, err := json.Marshal(payload)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
	c.Data(http.StatusOK, "application/json", jsonBytes)
}

func serveSwaggerUI(c *gin.Context) {
	html := strings.ReplaceAll(swaggerHTML, "__SPEC_URL__", c.Request.URL.Scheme+"://"+c.Request.Host+"/openapi.json")
	if c.Request.URL.Scheme == "" {
		html = strings.ReplaceAll(swaggerHTML, "__SPEC_URL__", "/openapi.json")
	}
	c.Data(http.StatusOK, "text/html; charset=utf-8", []byte(html))
}

const swaggerHTML = `<!DOCTYPE html>
<html lang="en">
	<head>
		<meta charset="UTF-8" />
		<title>Petstore API Docs</title>
		<link rel="stylesheet" href="https://unpkg.com/swagger-ui-dist@5/swagger-ui.css" />
		<style>
			body { margin: 0; background: #fafafa; }
			#swagger-ui { width: 100%; height: 100vh; }
		</style>
	</head>
	<body>
		<div id="swagger-ui"></div>
		<script src="https://unpkg.com/swagger-ui-dist@5/swagger-ui-bundle.js" crossorigin></script>
		<script>
			window.onload = () => {
				window.ui = SwaggerUIBundle({
					url: '__SPEC_URL__',
					dom_id: '#swagger-ui',
				});
			};
		</script>
	</body>
</html>`

type ApiHandleFunctions struct {

	// Routes for the PetAPI part of the API
	PetAPI PetAPI
	// Routes for the StoreAPI part of the API
	StoreAPI StoreAPI
	// Routes for the UserAPI part of the API
	UserAPI UserAPI
}

func getRoutes(handleFunctions ApiHandleFunctions) []Route {
	return []Route{
		{
			"AddPet",
			http.MethodPost,
			"/v2/pet",
			handleFunctions.PetAPI.AddPet,
		},
		{
			"DeletePet",
			http.MethodDelete,
			"/v2/pet/:petId",
			handleFunctions.PetAPI.DeletePet,
		},
		{
			"FindPetsByStatus",
			http.MethodGet,
			"/v2/pet/findByStatus",
			handleFunctions.PetAPI.FindPetsByStatus,
		},
		{
			"FindPetsByTags",
			http.MethodGet,
			"/v2/pet/findByTags",
			handleFunctions.PetAPI.FindPetsByTags,
		},
		{
			"GetPetById",
			http.MethodGet,
			"/v2/pet/:petId",
			handleFunctions.PetAPI.GetPetById,
		},
		{
			"UpdatePet",
			http.MethodPut,
			"/v2/pet",
			handleFunctions.PetAPI.UpdatePet,
		},
		{
			"UpdatePetWithForm",
			http.MethodPost,
			"/v2/pet/:petId",
			handleFunctions.PetAPI.UpdatePetWithForm,
		},
		{
			"GroomPet",
			http.MethodPost,
			"/v2/pet/:petId/groom",
			handleFunctions.PetAPI.GroomPet,
		},
		{
			"UploadFile",
			http.MethodPost,
			"/v2/pet/:petId/uploadImage",
			handleFunctions.PetAPI.UploadFile,
		},
		{
			"DeleteOrder",
			http.MethodDelete,
			"/v2/store/order/:orderId",
			handleFunctions.StoreAPI.DeleteOrder,
		},
		{
			"GetInventory",
			http.MethodGet,
			"/v2/store/inventory",
			handleFunctions.StoreAPI.GetInventory,
		},
		{
			"GetOrderById",
			http.MethodGet,
			"/v2/store/order/:orderId",
			handleFunctions.StoreAPI.GetOrderById,
		},
		{
			"PlaceOrder",
			http.MethodPost,
			"/v2/store/order",
			handleFunctions.StoreAPI.PlaceOrder,
		},
		{
			"CreateUser",
			http.MethodPost,
			"/v2/user",
			handleFunctions.UserAPI.CreateUser,
		},
		{
			"CreateUsersWithArrayInput",
			http.MethodPost,
			"/v2/user/createWithArray",
			handleFunctions.UserAPI.CreateUsersWithArrayInput,
		},
		{
			"CreateUsersWithListInput",
			http.MethodPost,
			"/v2/user/createWithList",
			handleFunctions.UserAPI.CreateUsersWithListInput,
		},
		{
			"DeleteUser",
			http.MethodDelete,
			"/v2/user/:username",
			handleFunctions.UserAPI.DeleteUser,
		},
		{
			"GetUserByName",
			http.MethodGet,
			"/v2/user/:username",
			handleFunctions.UserAPI.GetUserByName,
		},
		{
			"LoginUser",
			http.MethodGet,
			"/v2/user/login",
			handleFunctions.UserAPI.LoginUser,
		},
		{
			"LogoutUser",
			http.MethodGet,
			"/v2/user/logout",
			handleFunctions.UserAPI.LogoutUser,
		},
		{
			"UpdateUser",
			http.MethodPut,
			"/v2/user/:username",
			handleFunctions.UserAPI.UpdateUser,
		},
	}
}